fileutils: add support for copying symlinks to Copypath
This was removed instead of being special cased in constabulary/gb#185,
because gb has no direct need for it (see constabulary/gb#184).  However
it's useful for gvt, as it needs to retain test fixtures intact (as vendor/
packages are testable) and needs to support symlink-based packages (since
the go build tool does).

Fixes #12